refactor: test-case.ts 가져오는 함수 분리 및 중복코드 제거 작업
